+++ packages/common/src/discussions/api/channels/channels.ts
@@ -1,0 +1,17 @@
+import { request } from "../request";
+import { IChannel, IPagedResponse, ISearchChannelsParams } from "../types";
+
+/**
+ * Search for Channels in the Discussions API.  Channels define the capabilities,
+ * permissions, and configuration for Discussion posts.
+ *
+ * @export
+ * @param {ISearchChannelsParams} options
+ * @return {*}  {Promise<IPagedResponse<IChannel>>}
+ */
+export function searchChannels(
+  options: ISearchChannelsParams
+): Promise<IPagedResponse<IChannel>> {
+  options.httpMethod = "GET";
+  return request(`/channels`, options);
+}