+++ packages/common/src/discussions/index.ts
@@ -3,3 +3,4 @@ export * from "./edit";
 export * from "./fetch";
 export * from "./constants";
 export * from "./utils";
+export * from "./api";