+++ packages/common/src/search/types/IHubCatalog.ts
@@ -53,6 +53,13 @@ export interface IHubCollection {
   targetEntity: EntityType;
 }
 
+export type EntityType =
+  | "item"
+  | "group"
+  | "user"
+  | "groupMember"
+  | "event"
+  | "channel";
 /**
  * @private
  *
@@ -63,8 +70,6 @@ export interface IHubCollection {
 export interface IHubCollectionPersistance extends IHubCollection {
   hidden?: boolean;
 }
-
-export type EntityType = "item" | "group" | "user" | "groupMember" | "event";
 
 /**
  * IQuery is the fundamental unit used to execute a search. By composing