+++ packages/common/src/urls/index.ts
@@ -20,6 +20,10 @@ export * from "./getGroupHomeUrl";
 export * from "./getUserHomeUrl";
 export * from "./get-campaign-url";
 export * from "./is-safe-redirect-url";
+export * from "./cacheBustUrl";
+// For some reason, if this is exported here, random tests
+// start failing. Resolved by moving to the root index
+// export * from "./getCardModelUrl";
 
 const MAP_OR_FEATURE_SERVER_URL_REGEX = /\/(map|feature)server/i;