+++ packages/common/test/core/HubItemEntity.test.ts
@@ -36,6 +36,10 @@ describe("HubItemEntity Class: ", () => {
       authentication: MOCK_AUTH,
       currentUser: {
         username: "casey",
+        groups: [
+          { id: "efView", capabilities: [] },
+          { id: "efUpdate", capabilities: ["updateitemcontrol"] },
+        ],
       } as unknown as PortalModule.IUser,
       portal: {
         name: "DC R&D Center",
@@ -74,19 +78,107 @@ describe("HubItemEntity Class: ", () => {
         authdCtxMgr.context
       );
     });
-    it("shares to group", async () => {
+    it("shares to view group", async () => {
       const shareSpy = spyOn(PortalModule, "shareItemWithGroup").and.callFake(
         () => {
           return Promise.resolve();
         }
       );
-      await harness.shareWithGroup("3ef");
+      await harness.shareWithGroup("efView");
       expect(shareSpy).toHaveBeenCalledTimes(1);
       // verify args
       expect(shareSpy).toHaveBeenCalledWith({
         id: "00c",
-        groupId: "3ef",
+        groupId: "efView",
         owner: "deke",
+        confirmItemControl: false,
+        authentication: authdCtxMgr.context.session,
+      });
+    });
+    it("shares to update group", async () => {
+      const shareSpy = spyOn(PortalModule, "shareItemWithGroup").and.callFake(
+        () => {
+          return Promise.resolve();
+        }
+      );
+      await harness.shareWithGroup("efUpdate");
+      expect(shareSpy).toHaveBeenCalledTimes(1);
+      // verify args
+      expect(shareSpy).toHaveBeenCalledWith({
+        id: "00c",
+        groupId: "efUpdate",
+        owner: "deke",
+        confirmItemControl: true,
+        authentication: authdCtxMgr.context.session,
+      });
+    });
+
+    it("uses user groups to check if group is update", async () => {
+      const shareSpy = spyOn(PortalModule, "shareItemWithGroup").and.callFake(
+        () => {
+          return Promise.resolve();
+        }
+      );
+      await harness.shareWithGroup("notMyGroup");
+      expect(shareSpy).toHaveBeenCalledTimes(1);
+      // verify args
+      expect(shareSpy).toHaveBeenCalledWith({
+        id: "00c",
+        groupId: "notMyGroup",
+        owner: "deke",
+        confirmItemControl: false,
+        authentication: authdCtxMgr.context.session,
+      });
+    });
+    it("throws if user is not authd", async () => {
+      const unauthdCtxMgr = await ArcGISContextManager.create();
+      const instance = new TestHarness(
+        {
+          id: "00c",
+          owner: "deke",
+        },
+        unauthdCtxMgr.context
+      );
+      try {
+        await instance.shareWithGroup("efUpdate");
+        fail("should have thrown");
+      } catch (err) {
+        expect(err.name).toBe("HubError");
+      }
+    });
+    it("user w/o groups", async () => {
+      const ctxMgr = await ArcGISContextManager.create({
+        authentication: MOCK_AUTH,
+        currentUser: {
+          username: "casey",
+        } as unknown as PortalModule.IUser,
+        portal: {
+          name: "DC R&D Center",
+          id: "BRXFAKE",
+          urlKey: "fake-org",
+          customBaseUrl: "fakemaps.arcgis.com",
+        } as unknown as PortalModule.IPortal,
+      });
+      const instance = new TestHarness(
+        {
+          id: "00c",
+          owner: "deke",
+        },
+        ctxMgr.context
+      );
+      const shareSpy = spyOn(PortalModule, "shareItemWithGroup").and.callFake(
+        () => {
+          return Promise.resolve();
+        }
+      );
+      await instance.shareWithGroup("notMyGroup");
+      expect(shareSpy).toHaveBeenCalledTimes(1);
+      // verify args
+      expect(shareSpy).toHaveBeenCalledWith({
+        id: "00c",
+        groupId: "notMyGroup",
+        owner: "deke",
+        confirmItemControl: false,
         authentication: authdCtxMgr.context.session,
       });
     });
@@ -119,6 +211,7 @@ describe("HubItemEntity Class: ", () => {
       expect(setAccessSpy).toHaveBeenCalledWith({
         id: "00c",
         access: "public",
+        owner: "deke",
         authentication: authdCtxMgr.context.session,
       });
       // verify update to access property