+++ packages/common/test/initiatives/fixtures.ts
@@ -141,6 +141,7 @@ export const INITIATIVE_HUB_SEARCH_RESULT: IHubSearchResult = {
   updatedDate: new Date(1652819949000),
   updatedDateSource: "item.modified",
   family: "initiative",
+  index: 2,
   links: {
     self: "https://mock-home-url.com",
     siteRelative: "/mock-hub-relative-url",