+++ packages/common/test/search/_internal/getApi.test.ts
@@ -1,5 +1,5 @@
 import { IHubSearchOptions } from "../../../src/search/types/IHubSearchOptions";
-import { NamedApis } from "../../../src/search/types/types";
+import { IApiDefinition, NamedApis } from "../../../src/search/types/types";
 import { SEARCH_APIS } from "../../../src/search/utils";
 import { getApi } from "../../../src/search/_internal/commonHelpers/getApi";
 
@@ -30,6 +30,18 @@ describe("getApi", () => {
       type: "arcgis-hub",
       url: `${hubApiUrl}/api/search/v1`,
     });
+  });
+  it("otherwise returns reference to Discussions API if possible", () => {
+    const options = {
+      requestOptions: {
+        hubApiUrl,
+        isPortal: false,
+      },
+    } as unknown as IHubSearchOptions;
+    expect(getApi("channel", options)).toEqual({
+      type: "arcgis-hub",
+      url: null,
+    } as any as IApiDefinition);
   });
   it("otherwise returns a reference to the Portal API from requestOptions", () => {
     const portal = "https://my-enterprise-server.com/sharing/rest";